requirements.txt
# Main dependencies
python-dotenv==1.0.1
langchain==0.3.0
langchain-openai==0.2.0
svglib==1.5.1
langchain-experimental==0.3.1
cadquery==2.4.0
rlpycairo==0.3.0
loguru==0.7.2
langchainhub==0.1.15
streamlit==1.37.1
numpy==1.26.4
langchain-anthropic==0.3.0
langchain-google-genai==2.0.7

# Optional extras (include if you want vertexai support)
langchain-google-vertexai==2.0.1

# Development dependencies
black==24.4.2
isort==5.13.2